+++ pyhidra/__init__.py
@@ -1,15 +1,15 @@
 
-__version__ = "1.2.0"
+__version__ = "1.3.0"
 
 # Expose API
-from .core import run_script, start, started, open_program
+from .core import debug_callback, run_script, start, started, open_program
 from .launcher import DeferredPyhidraLauncher, GuiPyhidraLauncher, HeadlessPyhidraLauncher
 from .script import get_current_interpreter
 from .version import ApplicationInfo, ExtensionDetails
 
 
 __all__ = [
-    "get_current_interpreter", "open_program", "run_script", "start", "started",
+    "debug_callback", "get_current_interpreter", "open_program", "run_script", "start", "started",
     "ApplicationInfo", "DeferredPyhidraLauncher", "ExtensionDetails",
     "GuiPyhidraLauncher", "HeadlessPyhidraLauncher"
 ]